requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
gspread
oauth2client
uvicorn
requests
fpdf2
python-multipart
